login pelo face e mostrando loja

+++ lib/models/product.dart
@@ -8,25 +8,33 @@ import 'package:uuid/uuid.dart';
 import 'item_size.dart';
 
 class Product extends ChangeNotifier {
-
-  Product({this.id, this.name, this.description, this.images, this.sizes}){
-   images = images ?? [];
-   sizes = sizes ?? [];
+  Product(
+      {this.id,
+      this.name,
+      this.description,
+      this.images,
+      this.sizes,
+      this.deleted = false}) {
+    images = images ?? [];
+    sizes = sizes ?? [];
   }
 
-  Product.fromDocument(DocumentSnapshot document){
+  Product.fromDocument(DocumentSnapshot document) {
     id = document.documentID;
     name = document['name'] as String;
     description = document['description'] as String;
     images = List<String>.from(document.data['images'] as List<dynamic>);
-    sizes = (document.data['sizes'] as List<dynamic> ?? []).map(
-            (s) => ItemSize.fromMap(s as Map<String, dynamic>)).toList();
+    deleted = (document.data["deleted"] ?? false) as bool;
+    sizes = (document.data['sizes'] as List<dynamic> ?? [])
+        .map((s) => ItemSize.fromMap(s as Map<String, dynamic>))
+        .toList();
   }
 
   final Firestore firestore = Firestore.instance;
   final FirebaseStorage storage = FirebaseStorage.instance;
 
   DocumentReference get firestoreRef => firestore.document('products/$id');
+
   StorageReference get storageRef => storage.ref().child('products').child(id);
 
   String id;
@@ -34,53 +42,57 @@ class Product extends ChangeNotifier {
   String description;
   List<String> images;
   List<ItemSize> sizes;
+  bool deleted;
 
   List<dynamic> newImages;
 
   bool _loading = false;
+
   bool get loading => _loading;
-  set loading(bool value){
+
+  set loading(bool value) {
     _loading = value;
     notifyListeners();
   }
 
   ItemSize _selectedSize;
+
   ItemSize get selectedSize => _selectedSize;
-  set selectedSize(ItemSize value){
+
+  set selectedSize(ItemSize value) {
     _selectedSize = value;
     notifyListeners();
   }
 
   int get totalStock {
     int stock = 0;
-    for(final size in sizes){
+    for (final size in sizes) {
       stock += size.stock;
     }
     return stock;
   }
 
   bool get hasStock {
-    return totalStock > 0;
+    return totalStock > 0 && !deleted;
   }
 
   num get basePrice {
     num lowest = double.infinity;
-    for(final size in sizes){
-      if(size.price < lowest && size.hasStock)
-        lowest = size.price;
+    for (final size in sizes) {
+      lowest = size.price;
     }
     return lowest;
   }
 
-  ItemSize findSize(String name){
+  ItemSize findSize(String name) {
     try {
       return sizes.firstWhere((s) => s.name == name);
-    } catch (e){
+    } catch (e) {
       return null;
     }
   }
 
-  List<Map<String, dynamic>> exportSizeList(){
+  List<Map<String, dynamic>> exportSizeList() {
     return sizes.map((size) => size.toMap()).toList();
   }
 
@@ -91,9 +103,10 @@ class Product extends ChangeNotifier {
       'name': name,
       'description': description,
       'sizes': exportSizeList(),
+      "deleted": deleted,
     };
 
-    if(id == null){
+    if (id == null) {
       final doc = await firestore.collection('products').add(data);
       id = doc.documentID;
     } else {
@@ -102,23 +115,24 @@ class Product extends ChangeNotifier {
 
     final List<String> updateImages = [];
 
-    for(final newImage in newImages){
-      if(images.contains(newImage)){
+    for (final newImage in newImages) {
+      if (images.contains(newImage)) {
         updateImages.add(newImage as String);
       } else {
-        final StorageUploadTask task = storageRef.child(Uuid().v1()).putFile(newImage as File);
+        final StorageUploadTask task =
+            storageRef.child(Uuid().v1()).putFile(newImage as File);
         final StorageTaskSnapshot snapshot = await task.onComplete;
         final String url = await snapshot.ref.getDownloadURL() as String;
         updateImages.add(url);
       }
     }
 
-    for(final image in images){
-      if(!newImages.contains(image)){
+    for (final image in images) {
+      if (!newImages.contains(image) && image.contains("firebase")) {
         try {
           final ref = await storage.getReferenceFromUrl(image);
           await ref.delete();
-        } catch (e){
+        } catch (e) {
           debugPrint('Falha ao deletar $image');
         }
       }
@@ -131,14 +145,19 @@ class Product extends ChangeNotifier {
     loading = false;
   }
 
-  Product clone(){
+  Product clone() {
     return Product(
       id: id,
       name: name,
       description: description,
       images: List.from(images),
       sizes: sizes.map((size) => size.clone()).toList(),
+      deleted: deleted,
     );
+  }
+
+  void delete() {
+    firestoreRef.updateData({"deleted": true});
   }
 
   @override
